requirements.txt
pandas
sodapy==2.2.0
sqlalchemy 
psycopg2-binary
pyarrow
retry
tenacity
argparse
boto3